fix base type for shift

+++ clients/drcachesim/common/utils.h
@@ -151,7 +151,7 @@ compute_log2(int64_t value)
 {
     int i;
     for (i = 0; i < 63; i++) {
-        if (value == 1l << i)
+        if (value == int64_t(1) << i)
             return i;
     }
     // returns -1 if value is not a power of 2.
